Play with module structure.
