+++ lalrpop/src/lr1/backtrace/mod.rs
@@ -3,9 +3,11 @@ use lr1::{LR0Item, Item, State, StateIndex};
 use grammar::repr::*;
 use session::Session;
 
+mod example;
 mod state_graph;
 mod test;
 
+use self::example::ExampleIterator;
 use self::state_graph::StateGraph;
 
 pub struct Tracer<'trace, 'grammar: 'trace> {
@@ -40,6 +42,34 @@ pub struct BacktraceNode<'grammar> {
     parents: Vec<BacktraceNode<'grammar>>,
 }
 
+/// An "example" input and the way it was derived. This can be
+/// serialized into useful text. For example, it might represent
+/// something like this:
+///
+/// ```
+/// Ty "->" Ty "->" Ty
+/// |        |       |
+/// +-Ty-----+       |
+/// |                |
+/// +-Ty-------------+
+/// ```
+///
+/// The top-line is the `symbols` vector. The groupings below are
+/// stored in the `reductions` vector, in order from smallest to
+/// largest (they are always properly nested).
+#[derive(Debug)]
+pub struct Example {
+    pub symbols: Vec<Symbol>,
+    pub reductions: Vec<Reduction>,
+}
+
+#[derive(Debug)]
+pub struct Reduction {
+    pub start: usize,
+    pub end: usize,
+    pub nonterminal: NonterminalString,
+}
+
 impl<'grammar> BacktraceNode<'grammar> {
     fn new(item: Item<'grammar>) -> Self {
         BacktraceNode { item: LR0Item { production: item.production,
@@ -58,6 +88,10 @@ impl<'grammar> BacktraceNode<'grammar> {
         }
 
         self.parents.push(new_parent);
+    }
+
+    pub fn examples<'ex>(&'ex self) -> ExampleIterator<'ex> {
+        ExampleIterator::new(self)
     }
 }
 
@@ -149,141 +183,6 @@ impl<'trace, 'grammar> Tracer<'trace, 'grammar> {
         self.stack.pop().unwrap();
 
         result_node
-    }
-}
-
-pub fn examples<'ex>(backtrace_node: &'ex BacktraceNode<'ex>)
-                     -> ExampleIterator<'ex>
-{
-    ExampleIterator::new(backtrace_node)
-}
-
-pub struct ExampleIterator<'ex> {
-    stack: Vec<ExampleState<'ex>>,
-}
-
-#[derive(Debug)]
-pub struct Example {
-    pub symbols: Vec<Symbol>,
-    pub reductions: Vec<Reduction>,
-}
-
-#[derive(Debug)]
-pub struct Reduction {
-    pub start: usize,
-    pub end: usize,
-    pub nonterminal: NonterminalString,
-}
-
-#[derive(Debug)]
-struct ExampleState<'ex> {
-    // Node we are exploring
-    node: &'ex BacktraceNode<'ex>,
-
-    // Index of next parent to explore
-    index: usize,
-}
-
-impl<'ex> ExampleIterator<'ex> {
-    pub fn new(backtrace: &'ex BacktraceNode<'ex>) -> Self {
-        let mut this = ExampleIterator { stack: vec![] };
-        this.stack.push(ExampleState { node: backtrace, index: 0 });
-        this.populate();
-        this
-    }
-
-    fn populate(&mut self) -> bool {
-        println!("populate(self.stack={:?})", self.stack);
-        let parent = {
-            // Obtain parent from top of stack, if any, and increment
-            // index for top of stack.
-            let top = self.stack.last_mut().expect("populate called but stack is empty");
-            let index = top.index;
-            if index == top.node.parents.len() {
-                return false; // top of stack has no parent
-            }
-            top.index += 1;
-            &top.node.parents[index]
-        };
-        self.stack.push(ExampleState { node: parent, index: 0 });
-        self.populate();
-        return true; // top of stack had a parent (now pushed)
-    }
-
-    fn iterate(&mut self) {
-        // When this function is called, the top of the stack should
-        // always be some leaf node in the tree.
-        let top = self.stack.pop().unwrap();
-        assert!(top.node.parents.len() == 0 && top.index == 0);
-
-        println!("iterate(top={:?})", top);
-
-        while !self.stack.is_empty() {
-            if self.populate() {
-                return;
-            }
-
-            self.stack.pop();
-        }
-    }
-
-    fn unwind<I: Iterator<Item=&'ex LR0Item<'ex>>>(&self,
-                                                   mut rev_items: I,
-                                                   example: &mut Example) {
-        let item = if let Some(item) = rev_items.next() {
-            item
-        } else {
-            return;
-        };
-
-        println!("unwind(item={:?}", item);
-
-        let start = example.symbols.len();
-
-        let prefix = &item.production.symbols[..item.index];
-        println!("unwind: prefix={:?}", prefix);
-        example.symbols.extend(prefix);
-
-        self.unwind(rev_items, example);
-
-        if item.index != item.production.symbols.len() {
-            let suffix = &item.production.symbols[item.index+1..];
-            println!("unwind: suffix={:?}", suffix);
-            example.symbols.extend(suffix);
-        }
-
-        let end = example.symbols.len();
-        example.reductions.push(Reduction {
-            start: start,
-            end: end,
-            nonterminal: item.production.nonterminal
-        });
-    }
-}
-
-impl<'ex> Iterator for ExampleIterator<'ex> {
-    type Item = Example;
-
-    fn next(&mut self) -> Option<Example> {
-        println!("stack: {:?}", self.stack);
-
-        if self.stack.is_empty() {
-            return None;
-        }
-
-        let mut example = Example {
-            symbols: vec![],
-            reductions: vec![],
-        };
-
-        {
-            let rev_items = self.stack.iter().rev().map(|s| &s.node.item);
-            self.unwind(rev_items, &mut example);
-        }
-
-        self.iterate();
-
-        Some(example)
     }
 }
